# src/swagger_files/QuickCode.DemoCms.ContentManagementModule.Api-Swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "QuickCode - DemoCms - ContentManagementModule Api (1.0.0.0)",
    "termsOfService": "https://www.quickcode.net",
    "contact": {
      "name": "Üzeyir Apaydın",
      "url": "https://www.quickcode.net"
    },
    "license": {
      "name": "QuickCode.Net License",
      "url": "https://www.quickcode.net"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/content-management-module/pages/count": {
      "get": {
        "tags": [
          "Pages"
        ],
        "operationId": "PagesCount",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/content-management-module/pages/{id}": {
      "get": {
        "tags": [
          "Pages"
        ],
        "operationId": "PagesGetItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuickCode.DemoCms.ContentManagementModule.Application.Dtos.Page.PageDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Pages"
        ],
        "operationId": "PagesUpdate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuickCode.DemoCms.ContentManagementModule.Application.Dtos.Page.PageDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/QuickCode.DemoCms.ContentManagementModule.Application.Dtos.Page.PageDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/QuickCode.DemoCms.ContentManagementModule.Application.Dtos.Page.PageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Pages"
        ],
        "operationId": "PagesDelete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "QuickCode.DemoCms.ContentManagementModule.Application.Dtos.Page.PageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "contentTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "parentPageId": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/QuickCode.DemoCms.ContentManagementModule.Domain.Enums.ContentStatus"
          },
          "publishDate": {
            "type": "string",
            "format": "date-time"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "updatedDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "QuickCode.DemoCms.ContentManagementModule.Domain.Enums.ContentStatus": {
        "enum": [
          "Draft",
          "PendingReview",
          "Published",
          "Archived"
        ],
        "type": "string",
        "x-enumNames": [
          "Draft",
          "PendingReview",
          "Published",
          "Archived"
        ],
        "x-enum-varnames": [
          "Draft",
          "PendingReview",
          "Published",
          "Archived"
        ],
        "x-enumDescriptions": [
          "Content is in a draft state and not visible",
          "Content is awaiting approval",
          "Content is live and publicly visible",
          "Content is no longer visible but retained"
        ]
      }
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "Please enter a valid Api Key",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "ApiKey": [ ]
    }
  ]
}